trunk/trunk/requirements.txt
aliyun-python-sdk-core==2.13.36
aliyun-python-sdk-kms==2.15.0
asgiref==3.4.1
attrs==22.1.0
autopep8==1.6.0
beautifulsoup4==4.11.1
certifi==2022.6.15
cffi==1.15.1
chardet==5.0.0
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
crcmod==1.7
cryptography==37.0.4
dingtalk-sdk==1.3.8
Django==3.2.14
django-easy-audit==1.3.3
django-filter==21.1
django-oss-storage==1.1.1
django-silk==4.2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==4.4.0
drf-jwt==1.19.2
drf-spectacular==0.23.1
drf-spectacular-sidecar==2022.8.1
drf-yasg==1.21.3
gprof2dot==2021.2.21
idna==3.3
importlib-metadata==4.8.3
inflection==0.5.1
itsdangerous==2.0.1
itypes==1.2.0
Jinja2==3.0.3
jmespath==0.10.0
jsonschema==3.2.0
MarkupSafe==2.0.1
mysqlclient==2.1.1
optionaldict==0.1.2
oss2==2.16.0
packaging==21.3
Pillow==8.4.0
pycodestyle==2.8.0
pycparser==2.21
pycryptodome==3.15.0
Pygments==2.12.0
PyJWT==1.7.1
PyMySQL==1.0.2
pyparsing==3.0.9
pyrsistent==0.18.0
python-dateutil==2.8.2
python-weixin==0.5.0
pytz==2022.1
PyYAML==6.0
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
sentry-sdk==1.8.0
simplejson==3.17.6
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
toml==0.10.2
typing_extensions==4.1.1
uritemplate==4.1.1
urllib3==1.26.11
wechatpy==1.8.18
xmltodict==0.13.0
zipp==3.6.0
